// src/app/claim-processing/manual-claim-processing/process-claim-medsupp-charge/model/screen-save-area.model.ts
import {ScreenChargeLine} from './screen-charge-line.model';

/**
 * Model class ScreenSaveArea
 * Path: screenbean/procclmmedsupptbchrg
 * Model: com::uhc::aarp::fox::domain::screenbean::procclmmedsupptbchrg::ScreenSaveArea
 * Legacy Mapping: SCREEN-SAVE-AREA
 */
export class ScreenSaveArea {
  filler1 = '';
  filler2 = '';
  screenIcdCodes = '';
  scrnIcdCode2 = '';
  scrnIcdCode3 = '';
  screenChargeLines: ScreenChargeLine[] = [];
  screenAfterTermPlanL = 0;
  screenAfterTermPlanA = '';
  screenAfterTermPlan = '';
  filler13 = '';
  screenExclPlan1L = 0;
  screenExclPlan1A = '';
  screenExclPlan1 = '';
  filler14 = '';
  screenExclPlan2L = 0;
  screenExclPlan2A = '';
  screenExclPlan2 = '';
  m29cmnd = '';
  m29mem1 = '';
  m29mem2 = '';
  m29mem3 = '';
  m29cno1 = '';
  m29cno2 = '';
  m29cno3 = '';
  m29cno4 = '';
  m29cno5 = '';
  m29cno6 = '';
  m29com1 = '';
  m29com2 = '';
  m29err1 = '';
  m29name = '';
  m29memn = '';
  m29assc = '';
  m29dob = '';
  m29pln3 = '';
  m29eff2 = '';
  m29trm1 = '';
  m29trm2 = '';
  m29pln1 = '';
  m29pln2 = '';
  m29trm3 = '';
  m29eff3 = '';
  m29eff1 = '';
}